feat: add contract and script tasks
